organize package and test navigation
