+++ BudgetInHand/Views/Home/HomeView.swift
@@ -1,0 +1,20 @@
+//
+//  HomeView.swift
+//  BudgetInHand
+//
+//  Created by Quentin Bona on 07/12/2022.
+//
+
+import SwiftUI
+
+struct HomeView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct HomeView_Previews: PreviewProvider {
+    static var previews: some View {
+        HomeView()
+    }
+}
