+++ BudgetInHand/Views/Profile/ProfileView.swift
@@ -1,0 +1,20 @@
+//
+//  ProfileView.swift
+//  BudgetInHand
+//
+//  Created by Quentin Bona on 07/12/2022.
+//
+
+import SwiftUI
+
+struct ProfileView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct ProfileView_Previews: PreviewProvider {
+    static var previews: some View {
+        ProfileView()
+    }
+}
