+++ BudgetInHand/Views/Settings/SettingsView.swift
@@ -1,0 +1,20 @@
+//
+//  SettingsView.swift
+//  BudgetInHand
+//
+//  Created by Quentin Bona on 07/12/2022.
+//
+
+import SwiftUI
+
+struct SettingsView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct SettingsView_Previews: PreviewProvider {
+    static var previews: some View {
+        SettingsView()
+    }
+}
